+++ app/src/main/java/com/andreirookie/moviesapp/db/LocalDbInterface.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.LiveData
 import com.andreirookie.moviesapp.data.Movie
 
 interface LocalDbInterface {
-    fun removeById(movieId: Long)
+    fun removeById(movieId: Int)
     fun add(movie: Movie)
     fun getSaved(): LiveData<List<Movie>>
